requirements.txt
pandas
plotly
numpy
requests
beautifulsoup4
google-analytics-data
flask